+++ packages/components/src/components/checkbox-wrapper/checkbox-wrapper-styles.ts
@@ -14,7 +14,6 @@ import {
   prefersColorSchemeDarkMediaQuery,
 } from '../../styles';
 import {
-  borderRadiusMedium,
   borderRadiusSmall,
   fontFamily,
   fontLineHeight,
@@ -24,6 +23,7 @@ import {
 } from '@porsche-design-system/utilities-v2';
 import { getFunctionalComponentLabelStyles } from '../common/label/label-styles';
 import { getFunctionalComponentStateMessageStyles } from '../common/state-message/state-message-styles';
+import { getFunctionalComponentLoadingMessageStyles } from '../common/loading-message/loading-message-styles';
 
 const getCheckedSVGBackgroundImage = (fill: string): string => {
   return getInlineSVGBackgroundImage(
@@ -88,11 +88,6 @@ export const getComponentCss = (
                 }),
               },
             }),
-            ...(!isDisabled && {
-              '&(input:focus)::before': {
-                borderRadius: borderRadiusMedium,
-              },
-            }),
           },
         })
       ),
@@ -140,5 +135,7 @@ export const getComponentCss = (
     ...getFunctionalComponentStateMessageStyles(theme, state, {
       gridColumn: '1/3',
     }),
+    // .loading
+    ...getFunctionalComponentLoadingMessageStyles(),
   });
 };