+++ packages/components/src/styles/link-button-pure-styles.ts
@@ -5,7 +5,7 @@ import type { AlignLabel, BreakpointCustomizable, LinkButtonIconName, TextSize, 
 import {
   addImportantToEachRule,
   colorSchemeStyles,
-  getInsetJssStyle,
+  getFocusJssStyle,
   getThemedColors,
   getTransition,
   hostHiddenStyles,
@@ -14,7 +14,6 @@ import {
 } from './';
 import {
   borderRadiusSmall,
-  borderWidthBase,
   fontLineHeight,
   frostedGlassStyle,
   spacingStaticXSmall,
@@ -22,18 +21,19 @@ import {
 } from '@porsche-design-system/utilities-v2';
 import { getFontSizeText } from './font-size-text-styles';
 
+// TODO: why not using getHiddenTextJssStyle()?
 // Needed for slotted anchor and hidden label, which then enlarges the hidden label to equal host size and indents the text to be visually hidden.
 const getVisibilityJssStyle: GetJssStyleFunction = (hideLabel: boolean): JssStyle => {
   return hideLabel
     ? {
         position: 'absolute',
-        ...getInsetJssStyle(),
+        inset: 0,
         whiteSpace: 'nowrap',
-        textIndent: '-999999px',
+        textIndent: '-999999px', // TODO: check if text-indent still works for RTL-mode in this case
       }
     : {
         position: 'relative',
-        ...getInsetJssStyle('auto'),
+        inset: 'auto',
         whiteSpace: 'inherit',
         textIndent: 0,
         zIndex: 1, // fix Firefox bug on :hover (#2583)
@@ -55,12 +55,11 @@ export const getLinkButtonPureStyles = (
   hasSlottedAnchor: boolean,
   theme: Theme
 ): Styles => {
-  const { primaryColor, disabledColor, hoverColor, focusColor } = getThemedColors(theme);
+  const { primaryColor, disabledColor, hoverColor } = getThemedColors(theme);
   const {
     primaryColor: primaryColorDark,
     disabledColor: disabledColorDark,
     hoverColor: hoverColorDark,
-    focusColor: focusColorDark,
   } = getThemedColors('dark');
   const hasIcon = hasVisibleIcon(icon, iconSource);
 
@@ -90,7 +89,6 @@ export const getLinkButtonPureStyles = (
       ...prefersColorSchemeDarkMediaQuery(theme, {
         color: isDisabledOrLoading ? disabledColorDark : primaryColorDark,
       }),
-      outline: 0,
       ...textSmallStyle,
       ...mergeDeep(
         buildResponsiveStyles(stretch, (stretchValue: boolean) => ({
@@ -130,19 +128,10 @@ export const getLinkButtonPureStyles = (
             }),
           },
         })),
-      ...(!hasSlottedAnchor && {
-        '&:focus::before': {
-          border: `${borderWidthBase} solid ${focusColor}`,
-          ...prefersColorSchemeDarkMediaQuery(theme, {
-            borderColor: focusColorDark,
-          }),
-        },
-        '&:not(:focus-visible)::before': {
-          border: 0,
-        },
-      }),
+      ...(!hasSlottedAnchor && getFocusJssStyle(theme, { pseudo: true, offset: '-2px' })),
     },
     label: {
+      transition: getTransition('opacity'),
       position: 'relative', // needed for hover state when icon="none" is set
     },
     ...(hasIcon && {